+++ public/index.html
@@ -41,6 +41,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script type="module" src="./bundle.js"></script>
   </body>
 </html>
